Write clearCanvas function and add animations
Wipe the canvas clean and show an animation of the cells being cleared
by a line that moves across the canvas.

To make this possible, the canvas is now managed by a grid variable
stored in memory that all functions can globally access to find
references to the associated div cells in the canvas.